Add resizeobserver checks for remaining charts
